working system with 404 403 view pages

+++ database/migrations/2023_05_10_111124_create_billings_table.php
@@ -23,7 +23,7 @@ return new class extends Migration
                   ->references('id')
                   ->on('quotations')
                   ->onDelete('cascade');
-            $table->decimal('due');
+            $table->double('due');
             $table->string('payment_status')->default('unpaid');
             $table->longText('receipt_image')->nullable();
             $table->string('month');
